Render applepay btn on PDP (#1145)
* fix: removed err in catch

* feat: override detail page

* fix: unit tests

* chore: move files

* fix: tests

* fix: render button on initial render of the product
