+++ src/cartridges/app_adyen_SFRA/cartridge/client/default/js/adyen_checkout/__tests__/renderGenericComponent.test.js
@@ -69,9 +69,11 @@ beforeEach(() => {
     countryCode: 'mocked_countrycode',
   };
   getPaymentMethods.mockReturnValue({
-    adyenConnectedTerminals: { uniqueTerminalIds: ['mocked_id'] },
-    imagePath: 'example.com',
-    adyenDescriptions: {},
+    json: jest.fn().mockReturnValue({
+      adyenConnectedTerminals: { uniqueTerminalIds: ['mocked_id'] },
+      imagePath: 'example.com',
+      adyenDescriptions: {},
+    }),
   });
 });
 describe('Render Generic Component', () => {
