+++ src/cartridges/app_adyen_SFRA/cartridge/client/default/js/adyen_checkout/renderGenericComponent.js
@@ -217,7 +217,8 @@ function setGiftCardContainerVisibility() {
 }
 
 export async function initializeCheckout() {
-  const paymentMethodsResponse = await getPaymentMethods();
+  const paymentMethods = await getPaymentMethods();
+  const paymentMethodsResponse = await paymentMethods.json();
   const giftCardsData = await fetchGiftCards();
 
   setCheckoutConfiguration(paymentMethodsResponse);
