+++ src/cartridges/app_adyen_SFRA/cartridge/client/default/js/amazonPayExpressPart2.js
@@ -79,7 +79,8 @@ function showAddressDetails(shopperDetails) {
 async function mountAmazonPayComponent() {
   try {
     const amazonPayNode = document.getElementById('amazon-container');
-    const data = await getPaymentMethods();
+    const paymentMethods = await getPaymentMethods();
+    const data = await paymentMethods.json();
     const paymentMethodsResponse = data?.AdyenPaymentMethods;
     const applicationInfo = data?.applicationInfo;
     const checkout = await AdyenCheckout({
