+++ src/cartridges/app_adyen_SFRA/cartridge/client/default/js/productDetail.js
@@ -1,0 +1,7 @@
+/* eslint-disable global-require */
+const processInclude = require('base/util');
+
+$(document).ready(() => {
+  processInclude(require('base/product/detail'));
+  processInclude(require('./product/detail'));
+});
