+++ src/cartridges/int_adyen_SFRA/cartridge/adyen/utils/adyenHelper.js
@@ -770,7 +770,8 @@ let adyenHelperObj = {
   // get the fraction digits based on the currency code used to convert amounts of currency for the Adyen Checkout API
   getFractionDigits(currencyCode) {
     let format;
-    switch (currencyCode) {
+    let currency = currencyCode || session.currency.currencyCode;
+    switch (currency) {
       case 'CVE':
       case 'DJF':
       case 'GNF':
